Integer overflow fix for bug#331840428 	modified:   test/core/end2end/fuzzers/network_input.cc

+++ test/core/end2end/fuzzers/network_input.cc
@@ -506,7 +506,7 @@ Duration ScheduleConnection(
             Duration::NanosecondsRoundUp(
                 (q.slices.Length() * event_engine->max_delay_write()).count()));
   }
-  delay += Duration::Milliseconds(network_input.connect_delay_ms() +
+  delay += Duration::Milliseconds(static_cast<int64_t>(network_input.connect_delay_ms()) +
                                   network_input.connect_timeout_ms());
   event_engine->RunAfterExactly(
       Duration::Milliseconds(network_input.connect_delay_ms()),
